- Pulled Leader Group Selection Onto A Different Page
Made it so that if the person is a Coach, Captain, or Pastor of multiple groups, the main group leader toolbox page will only show the leader group. When clicking into the leader group, the groups under the leader group will be shown. This is to prevent a bunch of groups taking up a lot of space on this main page for Coaches, Captains, and Pastors.