- Added Check-in Step to Filter People That Are Restricted
Added a check-in workflow action that removes people that are restricted from checking in according to the specified date person attribute.

If the person has the specified attribute and the date contained in that attribute is equal to or after today, the person is unable to check in. If the person does not have a date in the attribute or the date is in the past, the check-in is left unaffected.